Added fuel code to markup examples and vocabulary.
fuel-edit-word - Asks for a word to edit, with completion.
With prefix, only words visible in the current vocabulary are
offered.

fuel-add-help-word-template -   Adds a help template for word at point.

fuel-help-quote-region -   Quotes a region as a string. Created to quote as pasted example from the listener

fuel-help-quote-indent -   Quotes and indents a region, presumably pasted example code.